change default customs slide every, sample events in alert
Rather than use a very dense sliding window slide value of 5 seconds,
increase this to by default 60 and use a trigger with pane processing
firing. This reduces the number of processing windows that are required.

Also add event data as a side input view so it can be included with any
resulting alert.