explicitly specify geoip db load path in parser
Reverts a bunch of the changes made in 231146e, and modifies
initialization so a GeoIP database path must be explicitly specified.
This prevents undesirable behavior where the test resource will be
loaded if the default resource is not found.

Pipelines must now explicitly enable GeoIP resolution in the parser or
in ParserDoFn by setting the correct input option.