dont call withoutDefaults in client count combine
Per the Beam documentation, if the input PCollection is not windowed
using global windows, withoutDefaults() or asSingletonView() must be
called.

https://beam.apache.org/releases/javadoc/2.11.0/org/apache/beam/sdk/transforms/Combine.Globally.html

Since we are using asSingletonView, remove the call to withoutDefaults
here to attempt to mitigate rare exceptions that are generated when the
window is empty, such as:

Caused by: java.lang.IllegalArgumentException: Empty PCollection accessed as a singleton view.
Consider setting withDefault to provide a default value