Set Event timestamp as time event occurred where possible
Sets Event.timestamp as time event occurred where possible. It defaults
"timestamp" within Stackdriver before setting it based on type-specific
fields.

For OpenSSH, year is not included, so the timestamp of when the event
was received is used to infer year. This could yield sadness on or
around new years midnight.