dont handle timestamps in event filter
Currently the setting associated with using the event timestamp during
pipeline stage emission is set within the event filter.

Change this so the setting is associated with ParserDoFn and happens
early in the process, since it shouldn't be associated with event
filtering at all.

This also makes it simpler to enable/disable as it just becomes an input
option.