cfgtick generation
Adds initial support for generation of periodic configuration ticks from
CompositeInput.

The messages are currently limited to current pipeline runtime options,
but are intended to be expanded to allow transforms to describe
functionality as well for inclusion in these messages.

By default, the new functionality is disabled.

The HTTPRequest pipeline has been modified to demonstrate use of the
generator.