remove customs generic rate limit analyzer
Removes the generic customs rate limit analyzer, which will be revisited
in future revisions here.

Also remove various supporting classes that were only utilized in the
rate limit analyzer. This includes the event filter configuration
support, and the keying selectors. This added unneeded complexity to the
event filter and will be reworked.