don't verify ParserCfg in setEventFilter
Since this validation occurs in a set function, an exception can be
triggered even if the parser is correctly specified in the input JSON if
the parser field is deserialized after the filter.

An exception is also generated if a parsed read is called without the
configuration being set, so just use this for validation instead.