improve discarded event handling
Rather than silently dropping events that do not match filters, use a
tuple tag to group these events in a separate collection. In the current
implementation, the count of events that were not OK is simply logged in
the pipeline. This could be expanded to potentially write certain types
of failures out to GCS as an example.

This also adds a new flag to drop events that have a timestamp older
than the specified period in seconds.