improve handling of datastore errors
Some updates to improve how errors occurring due to an outage in
Datastore are handled.

In customs velocity analysis, instead of raising this exception and
causing the pipeline to halt, log this at the info level and give up on
this particular analysis step.

In HTTPRequest, if a whitelist check fails log this as an error
condition and continue processing (but give up on this particular
analysis step).

wip