reduce no alert/notify log level to info
Instead of logging no alerting/notification configuration for an
identity as an error condition, log it at the info level as there are
cases where an identity may be present but have no notification
preferences.

Add functions so these types of configurations are logged as warnings
during AuthProfile graph construction.