remove string usage in metadata
Related to #358

This removes all usage of strings as metadata keys in Java code. This
does not yet cover off the relationship with cloud functions, which
would be the next step.

What also needs to happen following this is a larger consolidation of
key names as there are a lot of duplicates and potentially unused
values, and we can probably reduce the total key count.