improve handling of null input value
The string null is considered valid JSON input, and when processed by
jackson will return null from readValue() and will not trigger any sort
of JSON processing exception.

This causes problems in certain parsers that expect a non-null result if
the JSON parsing method does not throw an exception.

This change adds a test case for this scenario and adjusts the impacted
parsers.

In addition,

- Avoid reallocation of JacksonFactory for each CloudTrail message.
- Make sure relevant data is present while processing the Stackdriver
  header before we begin setting hints in parser state.